Refactor vendor management function to avoid duplicate db calls
