+++ Application/EdFi.Ods.Admin.Api/Features/Connect/TokenService.cs
@@ -1,4 +1,4 @@
-﻿// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: Apache-2.0
 // Licensed to the Ed-Fi Alliance under one or more agreements.
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
@@ -6,6 +6,7 @@
 using System.Security.Authentication;
 using System.Security.Claims;
 using EdFi.Ods.Admin.Api.Infrastructure.Security;
+using EdFi.Ods.AdminApp.Management.ErrorHandling;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using OpenIddict.Abstractions;
