+++ Application/EdFi.Ods.AdminApp.Management/Database/Commands/AddVendorCommand.cs
@@ -1,4 +1,4 @@
-﻿// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: Apache-2.0
 // Licensed to the Ed-Fi Alliance under one or more agreements.
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
@@ -20,7 +20,7 @@ namespace EdFi.Ods.AdminApp.Management.Database.Commands
             _context = context;
         }
 
-        public int Execute(IAddVendorModel newVendor)
+        public Vendor Execute(IAddVendorModel newVendor)
         {
             var namespacePrefixes = new List<VendorNamespacePrefix>();
             if (!string.IsNullOrWhiteSpace(newVendor.NamespacePrefixes))
@@ -50,7 +50,7 @@ namespace EdFi.Ods.AdminApp.Management.Database.Commands
 
             _context.Vendors.Add(vendor);
             _context.SaveChanges();
-            return vendor.VendorId;
+            return vendor;
         }
     }
