+++ Application/EdFi.Ods.AdminApp.Management/Database/Commands/DeleteVendorCommand.cs
@@ -1,10 +1,13 @@
-﻿// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: Apache-2.0
 // Licensed to the Ed-Fi Alliance under one or more agreements.
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
+using System;
 using System.Linq;
 using EdFi.Admin.DataAccess.Contexts;
+using EdFi.Ods.AdminApp.Management.Database.Queries;
+using EdFi.Ods.AdminApp.Management.ErrorHandling;
 
 namespace EdFi.Ods.AdminApp.Management.Database.Commands
 {
@@ -24,7 +27,13 @@ namespace EdFi.Ods.AdminApp.Management.Database.Commands
             var vendor = _context.Vendors.SingleOrDefault(v => v.VendorId == id);
 
             if (vendor == null)
-                return;
+            {
+                throw new NotFoundException<int>("This vendor no longer exists.It may have been recently deleted.", id);
+            }
+            if (vendor.IsSystemReservedVendor())
+            {
+                throw new Exception("This Vendor is required for proper system function and may not be deleted");
+            }
 
             foreach (var application in vendor.Applications.ToList())
             {
