+++ Application/EdFi.Ods.AdminApp.Management/Database/Commands/EditVendorCommand.cs
@@ -1,12 +1,15 @@
-﻿// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: Apache-2.0
 // Licensed to the Ed-Fi Alliance under one or more agreements.
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using EdFi.Admin.DataAccess.Contexts;
 using EdFi.Admin.DataAccess.Models;
+using EdFi.Ods.AdminApp.Management.Database.Queries;
+using EdFi.Ods.AdminApp.Management.ErrorHandling;
 using VendorUser = EdFi.Admin.DataAccess.Models.User;
 
 namespace EdFi.Ods.AdminApp.Management.Database.Commands
@@ -20,9 +23,18 @@ namespace EdFi.Ods.AdminApp.Management.Database.Commands
             _context = context;
         }
 
-        public void Execute(IEditVendor changedVendorData)
+        public Vendor Execute(IEditVendor changedVendorData)
         {
             var vendor = _context.Vendors.Single(v => v.VendorId == changedVendorData.VendorId);
+            if(vendor == null)
+            {
+                throw new NotFoundException<int>("This vendor no longer exists.It may have been recently deleted.", changedVendorData.VendorId);
+            }
+            if (vendor.IsSystemReservedVendor())
+            {
+                throw new Exception("This vendor is required for proper system function and may not be modified.");
+            }
+
             vendor.VendorName = changedVendorData.Company;
 
             if (vendor.VendorNamespacePrefixes.Any())
@@ -65,6 +77,7 @@ namespace EdFi.Ods.AdminApp.Management.Database.Commands
             }
 
             _context.SaveChanges();
+            return vendor;
         }
     }
